+++ RELEASE_NOTES.md
@@ -12,6 +12,8 @@
 
 - If your replay file contains a `_extensions` key, you should remove it, as you most likely want to use the extensions declared by the repo-config cookiecutter template you are upgrading to, otherwise you could get errors about missing extensions.
 
+- If your replay file contains a long `Introduction` key, you can replace it with an empty string (`""`), it doesn't need to have any particular content and it increases the size and noise in the replay file.
+
 ## New Features
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->
@@ -30,6 +32,10 @@
   It is not needed in the generated project, we always want to use the ones from the repo-config template.
 
   This should ease upgrading projects, making it less likely to have errors about missing extensions.
+
+- Clean up the `Introduction` variable from the generated replay file.
+
+  This is just a hack to be able to show a help about the template variables, keeping that text only increases the size and noise in the replay file.
 
 ## Bug Fixes